Separate concerns into modules
Separate concerns into modules for better organization and testing.

* **gherkin_parser.py**: Add `parse_feature_file` and `check_formatting` functions from `Gherkin2Fibery.py`.
* **csv_writer.py**: Add `write_to_csv`, `extract_scenarios`, `write_scenarios`, and `write_features_and_steps` functions from `Gherkin2Fibery.py`.
* **main.py**: Add `main` function from `Gherkin2Fibery.py` and import functions from `gherkin_parser.py` and `csv_writer.py`.
* **Gherkin2Fibery.py**: Delete file as its content is now distributed among the new modules.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/Zeegar/FeatureFile2Fibery?shareId=XXXX-XXXX-XXXX-XXXX).
